# api/swagger.yaml
openapi: 3.0.1
info:
  title: Smart Contract API SeedyFiuba
  description: 'API de interaccion con el SmartContract de SeedyFiuba'
  version: 1.3.0
servers:
  - url: https://seedifyuba-taller2.herokuapp.com
  - url: http://localhost:5000
...
paths:
...
  /projects:
    post:
      tags:
        - Projects
      summary: Alta de Proyecto
      description: Crea un proyecto en el SC y lo ingresa en el sistema persistiendo toda su informacion.
      operationId: createProject
      requestBody:
        description: Datos del usuario propietario o emprendedor del proyecto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInputDTO'
        required: true
      responses:
        200:
          description: Transaccion para crear proyecto realizada correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDataDTO'
        400:
          description: Bad Request. Alguno de los campos ingresados en invalido o esta incompleto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        500:
          description: Ocurrio un error inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
      x-codegen-request-body-name: project
...
  /transfers:
    post:
      tags:
        - Transfers
      summary: Transferencia
      description: SeedyFiuba permite realizar transferencias a usuarios externos. Realiza una transferencia de una Wallet a otra. Recibe el emisor, receptor y el monto a transferir en MiliEthers. El emisor debe ser un usuario registrado, mientras que el receptor tambien puede ser un tercero (externo).
      operationId: transfer
      requestBody:
        description: Datos de la transferencia, emisor, receptor y monto en MiliEthers.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferDTO'
        required: true
      responses:
        200:
          description: Transaccion para transferir realizada correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDataDTO'
        400:
          description: Bad Request. Alguno de los campos ingresados en invalido o esta incompleto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        500:
          description: Ocurrio un error inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
      x-codegen-request-body-name: transfer
...
  /users/{user_id}/transactions/{hash}:
    get:
      tags:
        - Transactions
      summary: Transacciones
      description: Obtiene la transaccion de un usuario segun su hash
      operationId: getTransaction
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
        - in: path
          name: hash
          required: true
          schema:
            type: string
      responses:
        200:
          description: Transferencia obtenida correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDataDTO'
        400:
          description: Bad Request. Alguno de los campos ingresados en invalido o esta incompleto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        500:
          description: Ocurrio un error inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedErrorResponse'
components:
  schemas:
...
    ProjectInputDTO:
      type: object
      properties:
        ownerId:
          description: Id del Usuario propietario. Este es el emprendedor del proyecto.
          type: string
          example: 'emprendedor1'
        reviewerId:
          description: Id del Usuario revisor. Este es el veedor del proyecto.
          type: string
          example: 'veedor1'
        stagesCost:
          description: Listado de costos de las etapas de cada proyecto.
          type: array
          items:
            type: integer
...
    TransferDTO:
      properties:
        sender_id:
          description: User ID del Emisor. Debe ser un usuario registrado en SeedyFiuba
          type: string
          example: 'user_1'
        receiver_address:
          description: Address Receptor o Destinatario. Puede o no ser un usuario registrado en Seedy Fiuba.
          type: string
          example: '0xabcdefghi'
        amount:
          description: Monto a transferir
          type: string
    TransactionDataDTO:
      type: object
      properties:
        hash:
          type: string
        status:
          type: string
          enum: [SUCCESSFUL, FAILURE]
          example: 'SUCCESSFUL'
        message:
          type: string
          example: 'operacion exitosa'
        addresss:
          description: Address de quien realizó la transaccion.
          type: string
          example: '0x123567890'
        project_id:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time
        flow:
          enum: [IN, OUT]
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [OK, FAILURE]
          example: 'OK'
        code:
          type: string
          example: 'successful_operation'
        message:
          type: string
          example: 'Successful Operation'
    BadRequestResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            status:
              example: 'FAILURE'
            error:
              example: 'Bad Request'
            statusCode:
              example: 400
            message:
              example: 'body should have required property "id"'
...
    UnexpectedErrorResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            status:
              example: 'FAILURE'
            code:
              example: 'unexpected_error'
            statusCode:
              example: 500
            message:
              example: 'Unexpected Error!'